requirements.txt
Flask==2.0.1
google-api-python-client==2.18.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
    